instances optional_dipping_sauce and dark_soy_sauce were NOT merged together
Query: in the Sauce class, should the instances 'Dark soy sauce' and 'Optional dipping sauce' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No.

Dark soy sauce and optional dipping sauce do not refer to the same thing and therefore should not be merged as semantic and ontological duplicates. Dark soy sauce is a type of soy sauce that has been aged for a longer period of time, resulting in a thicker consistency and darker color than other types of soy sauce. It is used primarily for cooking and adding depth of flavor to dishes. Optional dipping sauce, on the other hand, refers to any type of sauce that can be served alongside food for dipping purposes. This could include various regional and cultural sauces such as tzatziki, hummus, salsa, aioli, etc., but it does not specifically refer to dark soy sauce.
meaning that instances optional_dipping_sauce and dark_soy_sauce are different.
Query cache in files:
- /kgfiller/data/cache-31a4391c9c8234e32f01e8b4dfdc1a48ac7123f0be1d69a90f59872adaa8483e.yml