instances optional_dipping_sauce and fish_sauce were NOT merged together
Query: in the Sauce class, should the instances 'Fish sauce' and 'Optional dipping sauce' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No.

While both "Fish sauce" and "Optional dipping sauce" might contain the word "sauce," they do not refer to the same type of entity and therefore are not semantic nor ontological duplicates. Fish sauce is a specific type of condiment made from fermented fish, while an optional dipping sauce could be any kind of sauce served on the side for dipping food into. Therefore, it would not be appropriate to merge these two instances in this case.
meaning that instances optional_dipping_sauce and fish_sauce are different.
Query cache in files:
- /kgfiller/data/cache-cf9ff57b155144bd03c0a355e80eb7ffb7b9c8814d56fc90a29fe2e22a748ca9.yml