instances onion_powder and curry_powder were NOT merged together
Query: in the Food Seasoning class, should the instances 'Curry Powder' and 'Onion Powder' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No.

While curry powder and onion powder are both types of seasonings, they serve different purposes and have unique flavors and uses in cooking. Therefore, it would not be appropriate to merge them as semantic or ontological duplicates.
meaning that instances onion_powder and curry_powder are different.
Query cache in files:
- /kgfiller/data/cache-7cf036e4c855cf466edec0c2ca275efa8ede30bbb3247d6e4525addc8460dcf8.yml