instances greek_seasoning and cajun_seasoning were NOT merged together
Query: in the Food Seasoning class, should the instances 'Cajun Seasoning' and 'Greek Seasoning' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No.

Although both Cajun seasoning and Greek seasoning are types of spice blends used to add flavor to dishes, they consist of different ingredients and are used in different culinary traditions. Therefore, it is not appropriate to merge them as semantic and ontological duplicates.
meaning that instances greek_seasoning and cajun_seasoning are different.
Query cache in files:
- /kgfiller/data/cache-2a73eef66fad211bcc2d84e9899cc96f968f6eda114af1ec4dcf4a78ca437137.yml