instances artificial_flavors and chili_flakes were NOT merged together
Query: in the Food Seasoning class, should the instances 'Artificial flavors' and 'chili flakes' be merged together as semantic and ontologic duplicates? yes or no answer only.
Answer:
	 No.

Artificial flavors and chili flakes are not semantic nor ontological duplicates. Artificial flavors refer to synthetic or artificial substances that mimic the taste of natural ingredients, while chili flakes are made from dried and crushed red chillies and used as a seasoning or condiment for their spicy flavor and aroma. They belong to different categories with distinct characteristics and uses.
meaning that instances artificial_flavors and chili_flakes are different.
Query cache in files:
- /kgfiller/data/cache-d4d8644528cf1137d613958ab306397f41d25b7b1d25abfd7d0473ac2a78d1e8.yml